+++ Code/Core/Foundation/filesystem.h
@@ -17,7 +17,7 @@ namespace fs
   FOUNDATION_API FileStream open_file(const char* path);
   FOUNDATION_API void close_file(FileStream* file_stream);
 
-  FOUNDATION_API check_return bool write_file(FileStream file_stream, const void* src, size_t size);
-  FOUNDATION_API check_return bool read_file(FileStream file_stream, void* dst, size_t size);
+  FOUNDATION_API check_return bool write_file(FileStream file_stream, const void* src, u64 size);
+  FOUNDATION_API check_return bool read_file(FileStream file_stream, void* dst, u64 size);
   FOUNDATION_API u64 get_file_size(FileStream file_stream);
 }
